feat: コメント permalink を /comment/[id] に分離し /item は story 専用＋308リダイレクト (#164)
