+++ src/routes/comment/[id]/+page.svelte
@@ -1,0 +1,6 @@
+<script lang="ts">
+	import ItemPage from '$lib/components/ItemPage.svelte';
+	let { data, form } = $props();
+</script>
+
+<ItemPage {data} {form} />
