+++ src/routes/item/[id]/+page.server.ts
@@ -1,8 +1,10 @@
-import type { PageServerLoad, Actions } from './$types';
-import { getDB, getStoryById, getCommentsByStoryId, getCommentById, getChildComments, getCommentVoteStates, getVoteState, hasFavorited, hasFlagged, hasHidden, getFlaggedItemIds, getPollOptions, getPollOptionsVoted } from '$lib/server/db';
-import { TWO_WEEKS_MS } from '$lib/ranking';
-import { error, fail, redirect } from '@sveltejs/kit';
+import type { PageServerLoad } from './$types';
+import { getDB, getStoryById, getCommentsByStoryId, getCommentVoteStates, getVoteState, hasFavorited, hasFlagged, hasHidden, getFlaggedItemIds, getPollOptions, getPollOptionsVoted } from '$lib/server/db';
+import { error, redirect } from '@sveltejs/kit';
+import { actions } from '$lib/server/itemActions';
 
+// /item/[id] は story 専用。コメント permalink は /comment/[id] に分離した（#164）。
+// story でない id（コメント id 等）は /comment/{id} へ 308 リダイレクトする。
 export const load: PageServerLoad = async ({ params, platform, locals }) => {
 	const db = getDB(platform);
 	const id = parseInt(params.id, 10);
@@ -13,323 +15,62 @@ export const load: PageServerLoad = async ({ params, platform, locals }) => {
 
 	const showdead = locals.user?.showdead === 1;
 
-	// Try story first
 	const story = await getStoryById(db, id);
-	if (story) {
-		const comments = await getCommentsByStoryId(db, id, locals.user?.id, showdead);
-
-		let storyVoted = false;
-		let storyFavorited = false;
-		let storyFlagged = false;
-		let storyHidden = false;
-		let commentVoteStates: Map<number, 'up' | 'down'> = new Map();
-		let flaggedCommentIds: Set<number> = new Set();
-
-		if (locals.user) {
-			const [storyVoteState, fav, cvs, flag, fcids, hidden] = await Promise.all([
-				getVoteState(db, locals.user.id, id, 'story'),
-				hasFavorited(db, locals.user.id, id),
-				getCommentVoteStates(db, locals.user.id, comments.map((c) => c.id)),
-				hasFlagged(db, locals.user.id, id, 'story'),
-				getFlaggedItemIds(db, locals.user.id, comments.map((c) => c.id), 'comment'),
-				hasHidden(db, locals.user.id, id)
-			]);
-			storyVoted = storyVoteState === 'up';
-			storyFavorited = fav;
-			storyFlagged = flag;
-			commentVoteStates = cvs;
-			flaggedCommentIds = fcids;
-			storyHidden = hidden;
-		}
-
-		// poll の場合は選択肢と投票状況も取得する。
-		let pollOptions: Awaited<ReturnType<typeof getPollOptions>> = [];
-		let pollVotedOptionIds: number[] = [];
-		if (story.type === 'poll') {
-			pollOptions = await getPollOptions(db, story.id);
-			if (locals.user) {
-				const voted = await getPollOptionsVoted(db, locals.user.id, story.id);
-				pollVotedOptionIds = Array.from(voted);
-			}
-		}
-
-		return {
-			mode: 'story' as const,
-			story,
-			comments,
-			storyVoted,
-			storyFavorited,
-			storyFlagged,
-			storyHidden,
-			commentVoteStates: Object.fromEntries(commentVoteStates),
-			flaggedCommentIds: Array.from(flaggedCommentIds),
-			pollOptions,
-			pollVotedOptionIds
-		};
+	if (!story) {
+		// story でなければコメント permalink とみなして /comment/{id} へ恒久転送する。
+		throw redirect(308, '/comment/' + id);
 	}
 
-	// Not a story — try comment
-	const comment = await getCommentById(db, id);
-	if (!comment) {
-		throw error(404, 'Not found');
-	}
+	const comments = await getCommentsByStoryId(db, id, locals.user?.id, showdead);
 
-	const parentStory = await getStoryById(db, comment.story_id);
-	if (!parentStory) {
-		throw error(404, 'Parent story not found');
-	}
-
-	const childComments = await getChildComments(db, comment.id, comment.story_id, locals.user?.id, showdead);
-
-	let commentVoted = false;
-	let commentFlagged = false;
+	let storyVoted = false;
+	let storyFavorited = false;
+	let storyFlagged = false;
+	let storyHidden = false;
 	let commentVoteStates: Map<number, 'up' | 'down'> = new Map();
 	let flaggedCommentIds: Set<number> = new Set();
 
 	if (locals.user) {
-		const allCommentIds = [comment.id, ...childComments.map((c) => c.id)];
-		[commentVoteStates, flaggedCommentIds, commentFlagged] = await Promise.all([
-			getCommentVoteStates(db, locals.user.id, allCommentIds),
-			getFlaggedItemIds(db, locals.user.id, allCommentIds, 'comment'),
-			hasFlagged(db, locals.user.id, comment.id, 'comment')
+		const [storyVoteState, fav, cvs, flag, fcids, hidden] = await Promise.all([
+			getVoteState(db, locals.user.id, id, 'story'),
+			hasFavorited(db, locals.user.id, id),
+			getCommentVoteStates(db, locals.user.id, comments.map((c) => c.id)),
+			hasFlagged(db, locals.user.id, id, 'story'),
+			getFlaggedItemIds(db, locals.user.id, comments.map((c) => c.id), 'comment'),
+			hasHidden(db, locals.user.id, id)
 		]);
-		commentVoted = commentVoteStates.get(comment.id) === 'up';
+		storyVoted = storyVoteState === 'up';
+		storyFavorited = fav;
+		storyFlagged = flag;
+		commentVoteStates = cvs;
+		flaggedCommentIds = fcids;
+		storyHidden = hidden;
+	}
+
+	// poll の場合は選択肢と投票状況も取得する。
+	let pollOptions: Awaited<ReturnType<typeof getPollOptions>> = [];
+	let pollVotedOptionIds: number[] = [];
+	if (story.type === 'poll') {
+		pollOptions = await getPollOptions(db, story.id);
+		if (locals.user) {
+			const voted = await getPollOptionsVoted(db, locals.user.id, story.id);
+			pollVotedOptionIds = Array.from(voted);
+		}
 	}
 
 	return {
-		mode: 'comment' as const,
-		targetComment: comment,
-		parentStory,
-		comments: childComments,
-		commentVoted,
-		commentFlagged,
+		mode: 'story' as const,
+		story,
+		comments,
+		storyVoted,
+		storyFavorited,
+		storyFlagged,
+		storyHidden,
 		commentVoteStates: Object.fromEntries(commentVoteStates),
 		flaggedCommentIds: Array.from(flaggedCommentIds),
-		pollOptions: [] as Awaited<ReturnType<typeof getPollOptions>>,
-		pollVotedOptionIds: [] as number[]
+		pollOptions,
+		pollVotedOptionIds
 	};
 };
 
-async function resolveStory(db: D1Database, itemId: number): Promise<{ id: number; created_at: string }> {
-	const story = await getStoryById(db, itemId);
-	if (story) return story;
-	const comment = await getCommentById(db, itemId);
-	if (comment) {
-		const parentStory = await getStoryById(db, comment.story_id);
-		if (parentStory) return parentStory;
-	}
-	throw error(404, 'Not found');
-}
-
-export const actions: Actions = {
-	comment: async ({ request, platform, locals, params }) => {
-		if (!locals.user) {
-			throw redirect(302, '/login');
-		}
-
-		const db = getDB(platform);
-		const formData = await request.formData();
-		const text = (formData.get('text') as string)?.trim();
-		const parentId = formData.get('parent_id') as string | null;
-		const itemId = parseInt(params.id, 10);
-
-		if (!text) {
-			return fail(400, { error: 'Comment text is required', errorFor: 'comment' });
-		}
-
-		// Rate limit: 2 minutes between comments
-		const lastComment = await db
-			.prepare('SELECT created_at FROM comments WHERE user_id = ? ORDER BY created_at DESC LIMIT 1')
-			.bind(locals.user.id)
-			.first<{ created_at: string }>();
-
-		if (lastComment) {
-			const elapsed = Date.now() - new Date(lastComment.created_at).getTime();
-			if (elapsed < 2 * 60 * 1000) {
-				return fail(429, { error: "You're posting too fast. Please slow down.", text, errorFor: 'comment' });
-			}
-		}
-
-		const story = await resolveStory(db, itemId);
-
-		const elapsed = Date.now() - new Date(story.created_at).getTime();
-		if (elapsed >= TWO_WEEKS_MS) {
-			return fail(403, { error: 'Thread is closed' });
-		}
-		const parentIdNum = parentId ? parseInt(parentId, 10) : null;
-
-		await db
-			.prepare(
-				'INSERT INTO comments (text, user_id, story_id, parent_id) VALUES (?, ?, ?, ?)'
-			)
-			.bind(text, locals.user.id, story.id, parentIdNum)
-			.run();
-
-		await db
-			.prepare('UPDATE stories SET comment_count = comment_count + 1 WHERE id = ?')
-			.bind(story.id)
-			.run();
-
-		return { success: true };
-	},
-
-	editStory: async ({ request, platform, locals, params }) => {
-		if (!locals.user) {
-			throw redirect(302, '/login');
-		}
-
-		const db = getDB(platform);
-		const storyId = parseInt(params.id, 10);
-		const story = await getStoryById(db, storyId);
-
-		if (!story) {
-			throw error(404, 'Story not found');
-		}
-
-		if (story.user_id !== locals.user.id) {
-			throw error(403, 'Cannot edit another user\'s story');
-		}
-
-		const elapsed = Date.now() - new Date(story.created_at).getTime();
-		if (elapsed >= 2 * 60 * 60 * 1000) {
-			return fail(400, { error: 'Editing window has expired (2 hours)' });
-		}
-
-		const formData = await request.formData();
-		const title = (formData.get('title') as string)?.trim();
-		const text = (formData.get('text') as string) ?? '';
-
-		if (!title) {
-			return fail(400, { error: 'Title is required' });
-		}
-
-		// poll の編集では type='poll' を維持する。タイトル先頭が "Ask HN:" / "Show HN:"
-		// に変わっても type を書き換えない（書き換えると poll_options への参照は残るが
-		// /polls 一覧や [poll] タグから外れて poll 機能が事実上消失するため）。
-		// poll 以外は従来どおり title から自動判定。
-		let type: string;
-		if (story.type === 'poll') {
-			type = 'poll';
-		} else if (title.startsWith('Ask HN:')) {
-			type = 'ask';
-		} else if (title.startsWith('Show HN:')) {
-			type = 'show';
-		} else {
-			type = 'story';
-		}
-
-		await db
-			.prepare('UPDATE stories SET title = ?, text = ?, type = ? WHERE id = ?')
-			.bind(title, text || null, type, storyId)
-			.run();
-
-		return { success: true };
-	},
-
-	editComment: async ({ request, platform, locals }) => {
-		if (!locals.user) {
-			throw redirect(302, '/login');
-		}
-
-		const db = getDB(platform);
-		const formData = await request.formData();
-		const commentId = parseInt(formData.get('comment_id') as string, 10);
-		const text = (formData.get('text') as string)?.trim();
-
-		if (!text) {
-			return fail(400, { error: 'Comment text is required' });
-		}
-
-		const comment = await getCommentById(db, commentId);
-		if (!comment) {
-			throw error(404, 'Comment not found');
-		}
-
-		if (comment.user_id !== locals.user.id) {
-			throw error(403, 'Cannot edit another user\'s comment');
-		}
-
-		const elapsed = Date.now() - new Date(comment.created_at).getTime();
-		if (elapsed >= 2 * 60 * 60 * 1000) {
-			return fail(400, { error: 'Editing window has expired (2 hours)' });
-		}
-
-		await db
-			.prepare('UPDATE comments SET text = ? WHERE id = ?')
-			.bind(text, commentId)
-			.run();
-
-		return { success: true };
-	},
-
-	deleteStory: async ({ platform, locals, params }) => {
-		if (!locals.user) {
-			throw redirect(302, '/login');
-		}
-
-		const db = getDB(platform);
-		const storyId = parseInt(params.id, 10);
-		const story = await getStoryById(db, storyId);
-
-		if (!story) {
-			throw error(404, 'Story not found');
-		}
-
-		if (story.user_id !== locals.user.id) {
-			throw error(403, "Cannot delete another user's story");
-		}
-
-		const elapsed = Date.now() - new Date(story.created_at).getTime();
-		if (elapsed >= 2 * 60 * 60 * 1000) {
-			return fail(400, { error: 'Cannot delete after 2 hours' });
-		}
-
-		// 既に削除済みなら DB 書き込みをスキップ（冪等）
-		if (story.title === '[deleted]') {
-			return { success: true };
-		}
-
-		await db
-			.prepare('UPDATE stories SET title = ?, url = ?, text = ? WHERE id = ?')
-			.bind('[deleted]', null, '[deleted]', storyId)
-			.run();
-
-		return { success: true };
-	},
-
-	deleteComment: async ({ request, platform, locals }) => {
-		if (!locals.user) {
-			throw redirect(302, '/login');
-		}
-
-		const db = getDB(platform);
-		const formData = await request.formData();
-		const commentId = parseInt(formData.get('comment_id') as string, 10);
-
-		const comment = await getCommentById(db, commentId);
-		if (!comment) {
-			throw error(404, 'Comment not found');
-		}
-
-		if (comment.user_id !== locals.user.id) {
-			throw error(403, "Cannot delete another user's comment");
-		}
-
-		const elapsed = Date.now() - new Date(comment.created_at).getTime();
-		if (elapsed >= 2 * 60 * 60 * 1000) {
-			return fail(400, { error: 'Cannot delete after 2 hours' });
-		}
-
-		// 既に削除済みなら DB 書き込みをスキップ（冪等）
-		if (comment.text === '[deleted]') {
-			return { success: true };
-		}
-
-		await db
-			.prepare('UPDATE comments SET text = ? WHERE id = ?')
-			.bind('[deleted]', commentId)
-			.run();
-
-		return { success: true };
-	}
-};
+export { actions };
